+++ episuite/__init__.py
@@ -1,4 +1,6 @@
-__version__ = "0.1.0"
+__version__ = "0.2.0"
+__appname__ = "Episuite"
+__author__ = "Christian S. Perone"
 
 from episuite import durations  # noqa: F401
 from episuite import icu  # noqa: F401