+++ tests/test_mobility.py
@@ -16,9 +16,17 @@ class TestFacebookSurvey:
         item = result[0]
         assert set(["country", "region",
                     "survey_date"]).issubset(item.keys())
-        
+
     def test_get_survey_range(self) -> None:
         c = facebook.FacebookSymptomSurvey()
         result = c.get_survey_range("Brazil", "Rio Grande do Sul",
                                     "20210101", "20210105")
         assert len(result) == 5
+
+
+class TestFacebookMovementRange:
+    def test_last_date_avail(self) -> None:
+        mrange = facebook.FacebookMovementRange()
+        last_date = mrange._get_last_date_available()
+        assert "data.humdata.org" in last_date.url
+        assert last_date.date is not None